backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
numpy==2.0.1
pandas==2.2.3
scipy==1.15.3
pysfa==0.8
pyfrontier==1.0.2
scikit-learn==1.6.1
statsmodels==0.14.4
pytest==8.4.1
pytest-asyncio==0.24.0
httpx==0.28.1